fix: evaluation of `ConstPlugin`
